+++ sdk/messaging/azservicebus/example_admin_client_test.go
@@ -31,18 +31,18 @@ func ExampleNewAdminClientWithConnectionString() {
 	exitOnError("Failed to create ServiceBusClient in example", err)
 }
 
-func ExampleAdminClient_CreateQueue() {
-	queueProperties, err := adminClient.CreateQueue(context.TODO(), "queue-name")
-	exitOnError("Failed to create queue", err)
+func ExampleAdminClient_AddQueue() {
+	queueProperties, err := adminClient.AddQueue(context.TODO(), "queue-name")
+	exitOnError("Failed to add queue", err)
 
 	fmt.Printf("Queue name: %s", queueProperties.Name)
 }
 
-func ExampleAdminClient_CreateQueueWithProperties() {
+func ExampleAdminClient_AddQueueWithProperties() {
 	lockDuration := time.Minute
 	maxDeliveryCount := int32(10)
 
-	queueProperties, err := adminClient.CreateQueueWithProperties(context.TODO(), &azservicebus.QueueProperties{
+	queueProperties, err := adminClient.AddQueueWithProperties(context.TODO(), &azservicebus.QueueProperties{
 		Name: "queue-name",
 
 		// some example properties
@@ -52,4 +52,28 @@ func ExampleAdminClient_CreateQueueWithProperties() {
 	exitOnError("Failed to create queue", err)
 
 	fmt.Printf("Queue name: %s", queueProperties.Name)
+}
+
+func ExampleAdminClient_ListQueues() {
+	queuePager := adminClient.ListQueues(nil)
+
+	for queuePager.NextPage(context.TODO()) {
+		for _, queue := range queuePager.PageResponse() {
+			fmt.Printf("Queue name: %s, max size in MB: %d", queue.Name, queue.MaxSizeInMegabytes)
+		}
+	}
+
+	exitOnError("Failed when listing queues", queuePager.Err())
+}
+
+func ExampleAdminClient_ListQueuesRuntimeProperties() {
+	queuePager := adminClient.ListQueuesRuntimeProperties(nil)
+
+	for queuePager.NextPage(context.TODO()) {
+		for _, queue := range queuePager.PageResponse() {
+			fmt.Printf("Queue name: %s, active messages: %d", queue.Name, queue.ActiveMessageCount)
+		}
+	}
+
+	exitOnError("Failed when listing queues runtime properties", queuePager.Err())
 }